+++ src/main/kotlin/no/nav/syfo/common/token/SystemTokenProvider.kt
@@ -1,0 +1,5 @@
+package no.nav.syfo.common.token
+
+public interface SystemTokenProvider {
+    public suspend fun getSystemToken(targetClientId: String): String?
+}